change position2rank to double, rename RankList to RankChainFromExpert

+++ math/src/main/java/proeqfa/math/rank/IPosition2Rank.java
@@ -5,6 +5,6 @@ package proeqfa.math.rank;
  */
 public interface IPosition2Rank {
 
-    int getRankForPosition(int position);
+    double getRankForPosition(int position);
     
 }
